move get_train_val_split function into new dataset utils module
